requirements.txt
Pillow>=10.0.0
Flask>=2.3.0
Werkzeug>=2.3.0
